% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrayToDf}
\alias{arrayToDf}
\title{Convert an array to a data frame.}
\usage{
arrayToDf(ar)
}
\arguments{
\item{ar}{An array.}
}
\value{
\code{data.frame. dim = c(prod(dim(ar)), length(dim(ar)) + 1)}.
\itemize{
  \item Rows: One row per element of the array. Rows are "entered" into the data frame in
    \href{https://en.wikipedia.org/wiki/Row-_and_column-major_order}{column major order}.
  \item Columns: One column per dimension of the array, plus one column for the values in the
    array. \itemize{
    \item Column names are retained from names of array dimensions (\code{names(dimnames(ar)}).
      Where non-existant, column names are given as "d#", where # is the index of the dimension.
    \item For each column, if the corresponding array dimension was named, then the column is of
      class \code{character}. Otherwise, the column is of class \code{numeric}.
    \item The last column (the values from the array) is named "value".
  }
}
}
\description{
Convert an array to a data frame.
}
\seealso{
\code{\link[tidyr]{pivot_longer}}, \url{https://stackoverflow.com/a/42810479}.
}
