% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfToArray}
\alias{dfToArray}
\title{Convert a data frame to an array.}
\usage{
dfToArray(df, dimOrders = NULL)
}
\arguments{
\item{df}{\code{data.frame. dim = c(nr, nc)}. The first (nc - 1) columns represent dimensions.
The last column gives values in the array.}

\item{dimOrders}{\code{list} of \code{vector}. Ordering (i.e., factor levels) of each dimension
in the array. Must be a fully named list (arbitrary length) or fully unnamed list (\code{length
= nc}). If a column ordering is given for a column, any values present in that column but not
in the ordering assumes a value of NA. Any columns for which an ordering is not given is sorted
lexicographically. The class of each vector should match the class of the corresponding column.}
}
\value{
\code{array. mode = mode(df[[length(df)]])}. Mode of array is the same as the mode of the
  values column (i.e., last column) in the data frame.
}
\description{
Convert a data frame to an array.
}
\examples{
dfToArray(CO2) # CO2 data frame exported from datasets package

}
\references{
\url{https://stackoverflow.com/a/9617424}, \url{https://stackoverflow.com/a/46129338}.
}
