% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_cpus}
\alias{get_available_cpus}
\title{Get the number of CPUs the current process can use.}
\usage{
get_available_cpus(silent = FALSE)
}
\arguments{
\item{silent}{\code{logical}. Warn if unable to accurately detect number of available CPUs.}
}
\value{
\code{integer} or \code{NA} if the number of CPUs in the system is undetermined.
}
\description{
Get the number of CPUs the current process can use.
}
\seealso{
\link[parallel]{mcaffinity}, \link[parallel]{detectCores}.
}
