% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indexArray}
\alias{indexArray}
\title{Select along one dimension in multidimensional array}
\usage{
indexArray(x, dim, value, drop = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{\code{array}.}

\item{dim}{\code{integer}. Dimension along which to select one hyperplane.}

\item{value}{\code{integer}. Hyperplane to select from given dimension.}

\item{drop}{\code{logical}. For matrices and arrays. If TRUE the result is coerced to the lowest
possible dimension.}

\item{verbose}{\code{logical}. Print out equivalent command using `[` operator.}
}
\value{
\code{array} or \code{vector}.
}
\description{
Select along one dimension in multidimensional array
}
\examples{
A <- array(data = 1:24, dim = c(2, 3, 4), dimnames = list(c("x1", "x2"), c("y1", "y2", "y3"), NULL))
identical(indexArray(A, 1, 1), A[1, , ])
identical(indexArray(A, 1, 1, drop = FALSE), A[1, , , drop = FALSE])
identical(indexArray(A, 1, 2), A[2, , ])
identical(indexArray(A, 2, 3), A[, 3, ])

}
\references{
\url{https://stackoverflow.com/a/14502298}.
}
