% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orderArray}
\alias{orderArray}
\title{Order each dimension of array by hierarchical clustering.}
\usage{
orderArray(
  ar,
  dims = NULL,
  metric = "cor",
  method = "complete",
  cor_use = "pairwise.complete.obs",
  cor_method = "pearson",
  return_hclust = FALSE
)
}
\arguments{
\item{ar}{\code{array}.}

\item{dims}{\code{vector, integer}. Dimensions to order. If NULL, all dims are ordered.}

\item{metric}{\code{character}. Distance metric by which to compare hyperplanes along dimension.
Hyperplanes are flattened into vectors for comparison. Support \code{"cor"} ((1 - r) / 2) or
any \link[stats]{dist} method.}

\item{method}{\code{character}. Any \link[stats]{hclust} agglomeration method.}

\item{cor_use}{\code{character}. Only applicable if \code{metric = "cor"}. \code{use} argument of
\link[stats]{cor}.}

\item{cor_method}{\code{character}. Only applicable if \code{metric = "cor"}. \code{method}
argument of \link[stats]{cor}.}

\item{return_hclust}{\code{logical}. Return list of hclust objects for ordered dimensions.}
}
\value{
\code{list. length = length(dim(ar))}. \itemize{
  \item If \code{return_hclust = FALSE}: Each element is a vector giving the permutation of the
    corresponding array dimension.
  \item If \code{return_hclust = TRUE}: Each element is an \code{hclust} object for the
    corresponding dimension, or NULL if that dimension was not ordered.
}
}
\description{
Order each dimension of array by hierarchical clustering.
}
\examples{
ar <- matrix(c(1, 1, 1, 2, 2, 2, 3, 4, 5), nrow = 3)
orderArray(ar)

}
\seealso{
\link[stats]{cor}, \link[stats]{dist}, \link[stats]{hclust}.
}
