% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Evaluate an object or expression while suppressing output.}
\usage{
quiet(x, quiet_messages = TRUE, quiet_warnings = TRUE)
}
\arguments{
\item{x}{An object or expression}

\item{quiet_messages}{logical. Suppress messages.}

\item{quiet_warnings}{logical. Suppress warnings.}
}
\description{
Evaluate an object or expression while suppressing output.
}
\references{
Based on Hadley Wickham's suggestion in
  \url{https://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html}.
}
